refactor tests + new assert_test
